Dashboard page cleanup - Turn dashboard page logic into a view model - Use mixins for different groups of functions - Move footer and add widget dialog into separate files
